refactor(textwrap): use wrapped text module directly
